==> Final_Result/group_1_result/3.2_LG_고객데이터_HandlingNA_결측치_방법2.ipynb <==
# 3.2 범주형 결측지 처리 방법

- 이 파일에서는 여러 결측치 방법을 통해 결측치를 채운 후 각각 새로운 파일로 저장
- 파일 저장위치: `../data/결측치_output/`
- 이 데이터를 사용하여 `4.1_LG_고객데이터_feature_engineering+데이터생성.ipynb` 파일에서 가공변수 추가, 모델링에 사용할 데이터 생성

## 데이터 불러오기

from tqdm.auto import tqdm
import warnings
warnings.filterwarnings(action='ignore')
import pandas as pd
import numpy as np
from feature_engine.imputation import CategoricalImputer
from feature_engine.imputation import RandomSampleImputer


pd.set_option('display.max_columns', None)

# !pip install feature_engine

df = pd.read_csv("../data/customer.csv")
df.drop(['Unnamed: 0'], axis = 1, inplace = True)

df.drop(['케어솔루션고객여부', '거주행정동명', 'label_y', 'label_x'], axis = 1, inplace = True)

# 날짜형으로 변환
df['매출일자(배송완료일자)'] = pd.to_datetime(df['매출일자(배송완료일자)'], format='%Y-%m-%d')
df['년도'] = df['매출일자(배송완료일자)'].dt.year
df['월'] = df['매출일자(배송완료일자)'].dt.month

# 보유멤버십포인트 결측치는 0으로 채우기
df.loc[df['보유멤버십포인트'].isnull(), '보유멤버십포인트'] = 0

# 멤버십가입경로 4개 그룹으로 나눔

def change_variable(x):
    if x in ['베스트샵', '하이마트', '백화점', '전자랜드', '홈플러스', '(주)이마트']:
        return '매장'
    elif x in ['신한카드', '제휴카드', '하나SK카드', '롯데카드', 'LG페이', '네이버페이', '페이코', '스마트월렛']:
        return '카드'
    elif x in ['디얼디어', '바른손', '모닝글로리']:
        return '기타'
    elif x in ['LGE회원통합', 'LGE앱', 'LGE홈페이지(모바일)', '모바일그룹', '얍_KT', '얍_LGU', 'LGE홈페이지']:
        return '홈페이지'
    
def change_variable2(x):
    if x in ['백화점']:
        return '1'
    elif x in ['할인점 기타', '이마트', '온라인']:
        return '2'
    elif x in ['홈플러스', '하이마트', '전자랜드']:
        return '3'
    elif x in ['B2B', '베스트샵', '인터넷/홈쇼핑']:
        return '4'

df['멤버십가입경로'] = df['멤버십가입경로'].apply(lambda x: change_variable(x))
df['최근구매채널'] = df['최근구매채널'].apply(lambda x: change_variable2(x))
df.head(1)

df['케어솔루션품목수'] = df['케어솔루션품목수'].apply(lambda x : int(x[0]))
df['케어십품목수'] = df['케어십품목수'].apply(lambda x : int(x[0]))
df['Only&Best품목수'] = df['Only&Best품목수'].apply(lambda x : int(x[0]))
df['건강관리가전품목수'] = df['건강관리가전품목수'].apply(lambda x : int(x[0]))

X = df.copy()

X = X.drop('고객ID', axis = 1)

df_customer_ID = df[['고객ID']]

## Missing Data Imputation - feature engine

### CategoricalImputer() - 최빈값

# ignore_format=True --> numeric도 가능하게 하는 기능
ci = CategoricalImputer(imputation_method='frequent', ignore_format=True) 
ci.fit(X)
df_new = ci.transform(X)

df_new.isnull().sum()

ci.get_params()

df_new['연령대'].value_counts()

df_new = pd.concat([df_customer_ID, df_new], axis = 1)
# df_new

# df_new.to_csv('../data/결측치_output/결측치_CategoricalImputer.csv', encoding='utf-8-sig')

### CategoricalImputer(imputation_method='missing', ignore_format=True) : impute with an arbitrary value
- missing값으로 대체

# ignore_format=True --> numeric도 가능하게 하는 기능
ci = CategoricalImputer(imputation_method='missing', ignore_format=True) 
ci.fit(X)
df_missing = ci.transform(X)

df_missing['거주광역명'].value_counts()

df_missing = pd.concat([df_customer_ID, df_missing], axis = 1)

# df_missing.to_csv('../data/결측치_output/결측치_ArbitaryImputer.csv', encoding='utf-8-sig')

### RandomSampleImputer() - 랜덤

rsi = RandomSampleImputer()
rsi.fit(X)
df_rsi = rsi.transform(X)

df_rsi = pd.concat([df_customer_ID, df_rsi], axis = 1)


# df_rsi.to_csv('../data/결측치_output/결측치_RandomSampleImputer.csv', encoding='utf-8-sig')

df_rsi.columns

## Scikit Learn 



### KNN Imputer

https://scikit-learn.org/stable/modules/impute.html#knnimpute

사이킷런에서 설명하고 있는 KNN 임퓨터 작동 방식

각 표본의 결측치값은 학습 셋에서 찾은 n_neighbors 가장 가까운 이웃의 평균값을 사용하여 대치된다. 누락되지 않은 피처가 모두 가까운 경우에 두 샘플이 서로 이웃이 된다. 기본적으로 결측값인 nan_euclidean_distances를 지원하는 유클리드 디스턴스 메트릭이 가장 가까운 이웃을 찾는데 사용된다. 

결론 : 원하는 인접 이웃 수의 가중 또는 가중 평균을 사용하여 결측값을 대치한다. 

https://www.analyticsvidhya.com/blog/2020/07/knnimputer-a-robust-way-to-impute-missing-values-using-scikit-learn/

카테고리를 int로 수정

#### KNN 으로 채워줄때 범주형을 label encoding을 통해 int로 변환 필요

df_concat = df[['고객ID', '매출일자(배송완료일자)']]

X2 = df.copy()

X2.isnull().sum()

X2 = X2.drop(['고객ID', '매출일자(배송완료일자)'], axis = 1)

X2.info()

# list(air_df['보유멤버십포인트'].unique())

label_list1 = [0,  '1만↓', '3만↓', '5만↓', '10만↓','20만↓', '30만↓', '30만↑']
label_dict1 = {value: index for index, value in enumerate(label_list1)}

label_list2 = ['0원', '1KK미만', '1KK이상3KK미만','3KK이상5KK미만',  '5KK이상10KK미만','10KK이상15KK미만',
 '15KK이상20KK미만','20KK이상25KK미만','25KK이상30KK미만', '30KK↑']

label_dict2 = {value: index for index, value in enumerate(label_list2)}

# list(air_df['포인트사용빈도'].unique())
label_list3 = ['미사용','1~2회', '3~4회', '5~6회','7~8회','9~10회','11~12회','13~14회','15~16회','17~18회','19~20회', '20회↑']

label_dict3 = {value: index for index, value in enumerate(label_list3)}

# list(air_df['아파트평형대'].unique())
label_list4 = ['10평대↓', '10평대', '20평대', '30평대', '40평대', '50평대', '60평대','70평대↑']

label_dict4 = {value: index for index, value in enumerate(label_list4)}

# list(air_df['아파트시세'].unique())
label_list5 = [ '1억↓', '1~2억', '3~4억', '5~6억', '7~8억', '9~10억', '11~15억', '16~29억','30억↑']

label_dict5 = {value: index for index, value in enumerate(label_list5)}


# air_df.isnull().sum()

# list(air_df['금액'].unique())
label_list6 = ['', '0만원 이상 100만원 이하', '100만원 이상 200만원 이하', '200만원 이상 300만원 이하', 
               '300만원 이상 400만원 이하', '400만원 이상 500만원 이하', '500만원 이상 600만원 이하',
       '600만원 이상 700만원 이하', '700만원 이상 800만원 이하', '800만원 이상 900만원 이하', '900만원 이상 1000만원 이하', '1000만원 이상']

label_dict6 = {value: index for index, value in enumerate(label_list6)}

X2[f'금액'] = X2[f'금액'].map(label_dict6)

# list(air_df['월'].unique())
label_list7 = ['', 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12]

label_dict7 = {value: index for index, value in enumerate(label_list7)}

X2[f'월'] = X2[f'월'].map(label_dict7)


X2['보유멤버십포인트'] = X2['보유멤버십포인트'].map(label_dict1)
X2['최근3년구매금액'] = X2['최근3년구매금액'].map(label_dict2)
X2['포인트사용빈도'] = X2['포인트사용빈도'].map(label_dict3)
X2['아파트평형대'] = X2['아파트평형대'].map(label_dict4)
X2['아파트시세'] = X2['아파트시세'].map(label_dict5)

X2.isnull().sum()

X2['최근구매채널'].unique()

X2[(X2['최근구매채널'] != '1') & (X2['최근구매채널'] != '2') & (X2['최근구매채널'] != '3') & (X2['최근구매채널'] != '4')]['최근구매채널']

def changetoInt(x):
    if x != None:
        x = int(x)
        return x
    else:
        return x

X2['최근구매채널'] = X2['최근구매채널'].apply(changetoInt)

from sklearn.preprocessing import LabelEncoder

# Initialize the LabelEncoder
label_product = LabelEncoder()
label_age = LabelEncoder()
label_house = LabelEncoder()
label_membership = LabelEncoder()
label_city = LabelEncoder()
label_district = LabelEncoder()


X2['타겟제품명'] = label_product.fit_transform(X2['타겟제품명'])
X2['연령대'] = label_age.fit_transform(X2['연령대'])
X2['주거형태'] = label_house.fit_transform(X2['주거형태'])
X2['멤버십가입경로'] = label_membership.fit_transform(X2['멤버십가입경로'])
X2['거주광역명'] = label_city.fit_transform(X2['거주광역명'])
X2['거주시군구명'] = label_district.fit_transform(X2['거주시군구명'])

# LabelEncoder시, NaN 값들도 함께 인코딩이 되어서 다시 결측치로 변환
X2.loc[X2['연령대'] == (X2['연령대'].nunique()-1), '연령대']= np.nan
X2.loc[X2['멤버십가입경로'] == (X2['멤버십가입경로'].nunique()-1), '멤버십가입경로']= np.nan
X2.loc[X2['거주광역명'] == (X2['거주광역명'].nunique()-1), '거주광역명']= np.nan
X2.loc[X2['거주시군구명'] == (X2['거주시군구명'].nunique()-1), '거주시군구명']= np.nan

#### KNNImputer() 결측치 처리

from sklearn.impute import KNNImputer

imputer_uniform = KNNImputer(n_neighbors=25, weights="uniform")
df_knn_uniform = imputer_uniform.fit_transform(X2)

df_knn_uniform

imputer_distance = KNNImputer(n_neighbors=25, weights="distance")
df_knn_distance = imputer_distance.fit_transform(X2)

to_df_knn_uniform = pd.DataFrame(df_knn_uniform, columns = imputer_uniform.feature_names_in_)
to_df_knn_uniform

to_df_knn_distance = pd.DataFrame(df_knn_distance, columns = imputer_distance.feature_names_in_)
to_df_knn_distance

to_df_knn_uniform.to_csv('../data/결측치_output/결측치_KNNImputer_uniform25.csv', encoding='utf-8-sig')
to_df_knn_distance.to_csv('../data/결측치_output/결측치_KNNImputer_distance25.csv', encoding='utf-8-sig')

to_df_knn_uniform['타겟제품명'] = to_df_knn_uniform['타겟제품명'].apply(lambda x: int(x))
to_df_knn_uniform['연령대'] = to_df_knn_uniform['연령대'].apply(lambda x: int(x))
to_df_knn_uniform['주거형태'] = to_df_knn_uniform['주거형태'].apply(lambda x: int(x))
to_df_knn_uniform['멤버십가입경로'] = to_df_knn_uniform['멤버십가입경로'].apply(lambda x: int(x))
to_df_knn_uniform['거주광역명'] = to_df_knn_uniform['거주광역명'].apply(lambda x: int(x))
to_df_knn_uniform['거주시군구명'] = to_df_knn_uniform['거주시군구명'].apply(lambda x: int(x))

to_df_knn_distance['타겟제품명'] = to_df_knn_distance['타겟제품명'].apply(lambda x: int(x))
to_df_knn_distance['연령대'] = to_df_knn_distance['연령대'].apply(lambda x: int(x))
to_df_knn_distance['주거형태'] = to_df_knn_distance['주거형태'].apply(lambda x: int(x))
to_df_knn_distance['멤버십가입경로'] = to_df_knn_distance['멤버십가입경로'].apply(lambda x: int(x))
to_df_knn_distance['거주광역명'] = to_df_knn_distance['거주광역명'].apply(lambda x: int(x))
to_df_knn_distance['거주시군구명'] = to_df_knn_distance['거주시군구명'].apply(lambda x: int(x))

to_df_knn_uniform['타겟제품명'] = label_product.inverse_transform(to_df_knn_uniform['타겟제품명'])
to_df_knn_uniform['연령대'] = label_age.inverse_transform(to_df_knn_uniform['연령대'])
to_df_knn_uniform['주거형태'] = label_house.inverse_transform(to_df_knn_uniform['주거형태'])
to_df_knn_uniform['멤버십가입경로'] = label_membership.inverse_transform(to_df_knn_uniform['멤버십가입경로'])
to_df_knn_uniform['거주광역명'] = label_city.inverse_transform(to_df_knn_uniform['거주광역명'])
to_df_knn_uniform['거주시군구명'] = label_district.inverse_transform(to_df_knn_uniform['거주시군구명'])

to_df_knn_distance['타겟제품명'] = label_product.inverse_transform(to_df_knn_distance['타겟제품명'])
to_df_knn_distance['연령대'] = label_age.inverse_transform(to_df_knn_distance['연령대'])
to_df_knn_distance['주거형태'] = label_house.inverse_transform(to_df_knn_distance['주거형태'])
to_df_knn_distance['멤버십가입경로'] = label_membership.inverse_transform(to_df_knn_distance['멤버십가입경로'])
to_df_knn_distance['거주광역명'] = label_city.inverse_transform(to_df_knn_distance['거주광역명'])
to_df_knn_distance['거주시군구명'] = label_district.inverse_transform(to_df_knn_distance['거주시군구명'])

df_knn_uniform_final = pd.concat([df_concat, to_df_knn_uniform], axis = 1)
df_knn_distance_final = pd.concat([df_concat, to_df_knn_distance], axis = 1)


df_knn_uniform_final

df_knn_uniform_final.to_csv('../data/결측치_output/결측치_KNNImputer_uniform25_final.csv', encoding='utf-8-sig')
df_knn_distance_final.to_csv('../data/결측치_output/결측치_KNNImputer_distance25_final.csv', encoding='utf-8-sig')

# df_knn2.to_csv('../data/결측치_KNNImpter_10.csv', encoding='utf-8-sig')
