==> Final_Result/group_1_result/4_LG_고객데이터_feature_engineering+데이터생성.ipynb <==
# 4.1 Feature Engineering & 최종 데이터 생성

- 각 카테고리별, 결측치 처리 방법별 모델링에 사용할 데이터 생성
- 이 데이터를 사용하여 `5_LG_고객데이터_결측치_방법론선택.ipynb` 파일에서 각 카테고리에 맞는 결측치 처리 방법 선택

import pandas as pd
from tqdm import tqdm
from datetime import datetime

pd.set_option('display.max_columns', None)


### 결측치 완료한 데이터 프레임으로 수정 필요
# df = pd.read_csv('../data/결측치_output/결측치_f1기준_v1.csv', index_col = 0)

# df = pd.read_csv('../data/결측치_output/결측치_CategoricalImputer.csv', index_col = 0)

# df = pd.read_csv('../data/결측치_output/결측치_RandomSampleImputer.csv', index_col = 0)

# df = pd.read_csv('../data/결측치_ArbitaryImputer.csv', index_col = 0)

# df = pd.read_csv('../data/결측치_output/결측치_ArbitaryImputer.csv', index_col = 0)
# df = pd.read_csv('../data/결측치_output/결측치_KNNImputer_distance10_final.csv', index_col = 0)

## 2023년에 구매한 이력은 제거

df.loc[df['년도'] == 2023, 'label_2023'] = 1
df.loc[df['년도'] != 2023, 'label_2023'] = 0

df_isbuy_2023 = df[['고객ID', 'label_2023']]
df_isbuy_2023 = df_isbuy_2023[df_isbuy_2023['label_2023'] == 1]
df_isbuy_2023.drop_duplicates("고객ID", inplace = True)
df_isbuy_2023.rename({'label_2023':'label_y'}, axis = 1, inplace = True)
df_isbuy_2023

df['label_2023'].value_counts()

# 꼭 돌려야함
df = df.loc[df['label_2023'] != 1]

## 제품군 별로 나누기

def split_category(x):
    if x in ['냉장고','김치냉장고', '정수기', '식기세척기', '전기레인지', '얼음정수기냉장고', '광파오븐']:
        new_c = '부엌'
    elif x in ['TV', 'PC', '모니터', 'AV', '휴대폰']:
        new_c = '디스플레이'
    elif x in ['세탁기', '청소기', '건조기', '스타일러', '미니워시', '워시타워', '로봇청소기']:
        new_c = '청소'
    elif x in ['뷰티기기', '안마의자']:
        new_c = '건강관리'
    elif x in ['에어컨', '공기청정기']:
        new_c = '에어솔루션'
    else:
        new_c = '기타'
    return new_c

df['제품카테고리'] = df['타겟제품명'].apply(lambda x: split_category(x))

df[df['제품카테고리'] == '기타']['타겟제품명'].unique()

## 가공 변수 추가

- df_crm에 가공변수를 추가 
- 마지막에 df_crm과 제품별로 나눠진 데이터를 merge 필요

df_crm = pd.read_csv('../data/master_crm.csv', encoding='cp949')

### 총구매빈도

# df.loc[df['년도'] == 2023, 'label_2023'] = 1
# df.loc[df['년도'] != 2023, 'label_2023'] = 0

df_crm

df_crm = df_crm.merge(df.groupby(['고객ID']).count()['label_2023'], on = "고객ID")
df_crm = df_crm.rename(columns = {"label_2023" : "총구매빈도"})
df_crm.head(3)

### 최대구매가격대

df['금액_중간값'] = df['금액'].replace(to_replace={"Missing": 0, "0만원 이상 100만원 이하":50 , "100만원 이상 200만원 이하":150, "200만원 이상 300만원 이하":250, "300만원 이상 400만원 이하":350, "400만원 이상 500만원 이하":450, "500만원 이상 600만원 이하":550, "600만원 이상 700만원 이하":650, "700만원 이상 800만원 이하":750, "800만원 이상 900만원 이하":850, "900만원 이상 1000만원 이하":950, "1000만원 이상":1000 })

# 고객별로 그룹화하 후 각 그룹에서 최대값 리턴
max_purchase_by_customer = df.groupby('고객ID')['금액_중간값'].max()

df_crm['최대구매가격대'] = max_purchase_by_customer.reset_index(drop = True)
df_crm.head(2)

### 가장 최근 구매 이후 경과일

df['매출일자(배송완료일자)'] = pd.to_datetime(df['매출일자(배송완료일자)'])
current_date = datetime.now()

# 각 고객 별로 최근 매출일자(배송완료일자)를 찾고 일수 차이 계산
customer_last_purchase_dates = df.groupby('고객ID')['매출일자(배송완료일자)'].max().reset_index()
customer_last_purchase_dates['최근매출일로부터경과일'] = (current_date - customer_last_purchase_dates['매출일자(배송완료일자)']).dt.days

customer_last_purchase_dates

df_crm = df_crm.merge(customer_last_purchase_dates[['고객ID', '매출일자(배송완료일자)', '최근매출일로부터경과일']], on = "고객ID")

### 가장 선호하는 구매 제품

# 각 고객ID 별로 제품 카테고리의 구매 횟수를 계산
customer_purchase_counts = df.groupby(['고객ID', '제품카테고리']).size().reset_index(name='구매횟수')

# 각 고객ID 별로 가장 높은 구매 횟수를 찾아 해당하는 제품 카테고리를 반환
max_purchase_categories = customer_purchase_counts.groupby('고객ID').apply(lambda x: x[x['구매횟수'] == x['구매횟수'].max()]['제품카테고리'].tolist())

max_purchase_categories

df_crm['주구매제품카테고리'] = max_purchase_categories.reset_index(drop = True)
df_crm.head(2)

# 방법론 1 (선택)
- Idea : 각 제품군 별로 특징이 다를 것이므로, 제품군 별로 구매가능성을 예측하는 데에 더 정확할 것
- 따라서, 제품 카테고리 별로 데이터프레임을 구분한 후, 각각 제품군 별로 모델링을 하여, 모델을 여러 개 생성
- 이때, Y_target은 각 제품군 구매여부(1, 0)가 됨.
    - 2023년의 구매이력과 매칭하여, 어떤 제품군 카테고리를 구매했는지 확인할 수 있으므로, 각 제품군 별 구매여부 추적이 가능함

기본컬럼 (23) : '고객ID', '멤버십고객여부', '케어솔루션품목수', '케어십가입여부',
       '케어십품목수', '최근3년구매금액', '최근구매채널', 'Only&Best품목수', '건강관리가전품목수', '연령대',
       '주거형태', '아파트시세', '아파트평형대', '포인트사용빈도', '보유멤버십포인트', '멤버십가입경로', '제휴카드보유여부',
       '프리미엄카드보유여부', '제휴가입여부', '멤버십앱설치여부', '거주광역명', '거주시군구명', 'label_y'
       
추가 (4의배수) : '년도', '월', '타겟제품명', '금액'

가공변수 (4) : 총구매빈도, 최대구매가격대, 주구매제품카테고리

모델링에서 제거 : '최근매출일로부터경과일', '고객ID', '케어솔루션고객여부', '거주행정동명'

## 제품별로 데이터프레임 생성

### 모든 데이터

df_isbuy_2023

df_y = df.merge(df_isbuy_2023, on = '고객ID', how = 'left')
df_y.loc[df_y['label_y'].isnull(), 'label_y'] = 0

df_y['label_y'].value_counts()

df_y['년도'].unique()

### Air Solution

air_df = df_y.loc[df_y['제품카테고리'] == '에어솔루션']

# air_df.loc[:, 'label_y'] = 0
# air_df.loc[(air_df['년도'] == 2023) & (air_df['제품카테고리'] == '에어솔루션'), 'label_y'] = 1

air_df['label_y'].value_counts()

# air_df

### Clean

clean_df = df_y.loc[df_y['제품카테고리'] == '청소']

clean_df['label_y'].value_counts()

### Kitchen

kitchen_df = df_y.loc[df_y['제품카테고리'] == '부엌']


kitchen_df['label_y'].value_counts()

### Display

display_df = df_y.loc[df_y['제품카테고리'] == '디스플레이']
# display_df.loc[:, 'label_y'] = 0
# display_df.loc[(display_df['년도'] == 2023) & (display_df['제품카테고리'] == '디스플레이'), 'label_y'] = 1

display_df['label_y'].value_counts()

### Health

health_df = df_y.loc[df_y['제품카테고리'] == '건강관리']

health_df['label_y'].value_counts()

### Signature

etc_df = df_y.loc[df_y['제품카테고리'] == '기타']

etc_df['label_y'].value_counts()

## 고객의 구매이력 데이터 통합 함수

def create_row_from_group(group, customer_id):

    new_row = {'고객ID': customer_id}
    for column in group.columns:
        new_row[column] = group.iloc[0][column]
    return new_row

def organize_purchases_per_customer(dataframe, column_list):
    
    # 단일 고객의 최대 구매 횟수 파악
    grouped_new = dataframe.groupby('고객ID')
    max_purchases_new = grouped_new.size().max()

    label_encoding_df = {column: [] for column in column_list}

    for i in range(1, max_purchases_new + 1):
        label_encoding_df[f'년도_{i}'] = []
        label_encoding_df[f'월_{i}'] = []
        label_encoding_df[f'타겟제품명_{i}'] = []
        label_encoding_df[f'금액_{i}'] = []

    for customer_id, group in tqdm(grouped_new):
        purchases = list(group[['년도', '월','타겟제품명', '금액']].itertuples(index=False, name=None))
        new_row_new = create_row_from_group(group, customer_id)

        # 새 행에 판매 날짜 및 구매 추가
        for i, (year, month, product_name, amount) in enumerate(purchases, 1):
            new_row_new[f'년도_{i}'] = year
            new_row_new[f'월_{i}'] = month
            new_row_new[f'타겟제품명_{i}'] = product_name
            new_row_new[f'금액_{i}'] = amount

        # 최대 구매수보다 적은 경우 나머지 행 채우기
        for j in range(len(purchases) + 1, max_purchases_new + 1):
            new_row_new[f'년도_{j}'] = None
            new_row_new[f'월_{j}'] = None
            new_row_new[f'타겟제품명_{j}'] = None
            new_row_new[f'금액_{j}'] = None

        for key in label_encoding_df:
            label_encoding_df[key].append(new_row_new.get(key, None))

    new_structured_df_new = pd.DataFrame(label_encoding_df)
    return new_structured_df_new


## 고객별 데이터 생성 : 고객 정보와 가공변수 포함된 df 

def final_df_maker(dataframe, col_list):
    df_new = organize_purchases_per_customer(dataframe, col_list)
    # 가공변수 추가
    df_new = df_new.merge(df_crm[['고객ID','총구매빈도', '최대구매가격대',
       '주구매제품카테고리']], on = '고객ID', how = 'left')
    return df_new

col_list = [
'고객ID','멤버십고객여부', '케어솔루션품목수', '케어십가입여부', '케어십품목수', '최근3년구매금액', '최근구매채널',
'Only&Best품목수', '건강관리가전품목수', '연령대', '주거형태', '아파트시세', '아파트평형대', '포인트사용빈도',
'보유멤버십포인트', '멤버십가입경로', '제휴카드보유여부', '프리미엄카드보유여부', '제휴가입여부', '멤버십앱설치여부',
'거주광역명', '거주시군구명', 'label_y']

etc_df_new = final_df_maker(etc_df, col_list)

health_df_new = final_df_maker(health_df, col_list)

clean_df_new = final_df_maker(clean_df, col_list)

air_df_new = final_df_maker(air_df, col_list)

kitchen_df_new = final_df_maker(kitchen_df, col_list)
display_df_new = final_df_maker(display_df, col_list)

all_df_new = final_df_maker(df_y, col_list)

# etc_df_new.to_csv('../data/etc/etc_df_최빈값_2023제거.csv', encoding='utf-8-sig')


# etc_df_new.to_csv('../data/etc/etc_df_최빈값_2023제거.csv', encoding='utf-8-sig')
# health_df_new.to_csv('../data/health/health_df_최빈값_2023제거.csv', encoding='utf-8-sig')
# air_df_new.to_csv('../data/air/air_df_최빈값_2023제거.csv', encoding='utf-8-sig')
# kitchen_df_new.to_csv('../data/kitchen/kitchen_df_최빈값_2023제거.csv', encoding='utf-8-sig')
# display_df_new.to_csv('../data/display/display_df_최빈값_2023제거.csv', encoding='utf-8-sig')
# clean_df_new.to_csv('../data/clean/clean_df_최빈값_2023제거.csv', encoding='utf-8-sig')


# all_df_new.to_csv('../data/all/all_df_최빈값_2023제거.csv', encoding='utf-8-sig')

## 미세먼지, 온도 추가

df_x = df_y.copy()

### 미세먼지

dust1 = pd.read_excel('../data/외부변수/20091120-20181231_미세먼지농도.xlsx')

dust2 = pd.read_excel('../data/외부변수/20190101-20230831_미세먼지농도.xlsx')

dust = pd.concat([dust1, dust2])

dust

mean_dust = dust[['일시', '일 미세먼지 농도(㎍/㎥)']].groupby('일시').mean().reset_index()

### 위의 일 미세먼지 농도가 결측치인 것을 '서울시 일별 평균 대기오염도 정보'를 평균냄으로써 결측치 보완 (이하 날짜)
 https://data.seoul.go.kr/dataList/OA-2218/S/1/datasetView.do
* 2019-10-08 : 23.044
* 2019-10-09 : 24.804
* 2022-02-15 : 35.063

# 누락된 미세먼지 농도 데이터 추가

mean_dust.append(pd.DataFrame({'일시': ['2019-10-08'], '일 미세먼지 농도(㎍/㎥)': [23.044]}), ignore_index=True)
mean_dust.append(pd.DataFrame({'일시': ['2019-10-09'], '일 미세먼지 농도(㎍/㎥)': [24.804]}), ignore_index=True)
mean_dust.append(pd.DataFrame({'일시': ['2022-02-15'], '일 미세먼지 농도(㎍/㎥)': [35.063]}), ignore_index=True)

mean_dust['일시'] = pd.to_datetime(mean_dust['일시'], format='%Y-%m-%d')
mean_dust.sort_values(by = "일시", inplace = True)
mean_dust

mean_dust

df_x['매출일자(배송완료일자)'] = pd.to_datetime(df_x['매출일자(배송완료일자)'], format='%Y-%m-%d')

df_x = df_x.merge(mean_dust, left_on = '매출일자(배송완료일자)', right_on = '일시', how = 'left')

### 온도

temperature1 = pd.read_excel('../data/외부변수/20091120-20181231_기온.xlsx')

temperature2 = pd.read_excel('../data/외부변수/20190101-20230831_기온.xlsx')

temperature = pd.concat([temperature1, temperature2])

mean_temperature = temperature[['일시', '평균기온(°C)']].groupby('일시').mean().reset_index()

mean_temperature['일시'] = pd.to_datetime(mean_temperature['일시'], format='%Y-%m-%d')
mean_temperature

df_x = df_x.merge(mean_temperature, left_on = '매출일자(배송완료일자)', right_on = '일시', how = 'left')

df_x = df_x.drop(['일시_x', '일시_y'], axis = 1)

df_x = df_x.rename({'일 미세먼지 농도(㎍/㎥)' : '미세먼지', '평균기온(°C)' : '기온'}, axis = 1)

df_x.head(2)

def organize_purchases_per_customer_add_variable(dataframe, column_list):
    
    # 단일 고객의 최대 구매 횟수 파악
    grouped_new = dataframe.groupby('고객ID')
    max_purchases_new = grouped_new.size().max()

    label_encoding_df = {column: [] for column in column_list}

    for i in range(1, max_purchases_new + 1):
        label_encoding_df[f'년도_{i}'] = []
        label_encoding_df[f'월_{i}'] = []
        label_encoding_df[f'타겟제품명_{i}'] = []
        label_encoding_df[f'금액_{i}'] = []
        label_encoding_df[f'미세먼지_{i}'] = []
        label_encoding_df[f'기온_{i}'] = []

    for customer_id, group in tqdm(grouped_new):
        purchases = list(group[['년도', '월','타겟제품명', '금액', '미세먼지', '기온']].itertuples(index=False, name=None))
        new_row_new = create_row_from_group(group, customer_id)

        # 새 행에 판매 날짜 및 구매 추가
        for i, (year, month, product_name, amount, dust, temperature) in enumerate(purchases, 1):
            new_row_new[f'년도_{i}'] = year
            new_row_new[f'월_{i}'] = month
            new_row_new[f'타겟제품명_{i}'] = product_name
            new_row_new[f'금액_{i}'] = amount
            new_row_new[f'미세먼지_{i}'] = dust
            new_row_new[f'기온_{i}'] = temperature

        # 최대 구매수보다 적은 경우 나머지 행 채우기
        for j in range(len(purchases) + 1, max_purchases_new + 1):
            new_row_new[f'년도_{j}'] = None
            new_row_new[f'월_{j}'] = None
            new_row_new[f'타겟제품명_{j}'] = None
            new_row_new[f'금액_{j}'] = None
            new_row_new[f'미세먼지_{i}'] = None
            new_row_new[f'기온_{i}'] = None

        for key in label_encoding_df:
            label_encoding_df[key].append(new_row_new.get(key, None))

    new_structured_df_new = pd.DataFrame(label_encoding_df)
    return new_structured_df_new


def final_df_maker_add_variable(dataframe, col_list):
    df_new = organize_purchases_per_customer_add_variable(dataframe, col_list)
    # 가공변수 추가
    df_new = df_new.merge(df_crm[['고객ID','총구매빈도', '최대구매가격대',
       '주구매제품카테고리']], on = '고객ID', how = 'left')
    return df_new

col_list = [
'고객ID','멤버십고객여부', '케어솔루션품목수', '케어십가입여부', '케어십품목수', '최근3년구매금액', '최근구매채널',
'Only&Best품목수', '건강관리가전품목수', '연령대', '주거형태', '아파트시세', '아파트평형대', '포인트사용빈도',
'보유멤버십포인트', '멤버십가입경로', '제휴카드보유여부', '프리미엄카드보유여부', '제휴가입여부', '멤버십앱설치여부',
'거주광역명', '거주시군구명', 'label_y']

### Air Solution - 미세먼지, 기온

air2_df = df_x.loc[df_x['제품카테고리'] == '에어솔루션']

air2_df_new = final_df_maker_add_variable(air2_df, col_list)

air2_df_new.head(2)

air2_df_new.to_csv('../data/air_외부변수/air_df_new_값생성_미세먼지_2023제거.csv', encoding='utf-8-sig')

# 방법론 2 (사용하지 않음)
- Idea : 각 제품군 별로 방법론 1처럼 데이터프레임을 나누게 되면, 어떤 제품을 구매하고, 다른 제품군도 같이 구매하게 되는 제품 간 연관성 정보가 소실될 수 있음
- 따라서, 제품군 별로 데이터프레임을 구분하지 않고, 하나의 기존 데이터프레임으로 둠.
- 이때, Y_target은 제품군 별 구매여부(1, 0)를 매번 모델링 때마다 다르게 변화시키면서 모델 예측을 진행 ex) 동일한 데이터프레임을 가지고, 처음에는 Y_target을 부엌, 그 다음은 에어솔루션 등으로 정답지를 바꾸면서 각각 다른 모델을 도출하는 것

df2 = pd.read_csv('../data/.....', index_col = 0)
df_copy = df2.copy()

## 2023년 구매 여부 컬럼

# 데이터를 '년도' 기준으로 내림차순 정렬
df_copy = df_copy.sort_values(by='년도', ascending=False)

# 각 '고객ID'별로 첫 번째 행만 남기고 중복 제거
df_copy = df_copy.drop_duplicates(subset='고객ID')

df_copy.loc[df_copy['년도'] == 2023, '2023년구매'] = 1
df_copy.loc[df_copy['년도'] != 2023, '2023년구매'] = 0

df2 = df2.merge(df_copy[['고객ID', '2023년구매']], on='고객ID', how = 'left')

df2['제품카테고리'] = df2['타겟제품명'].apply(lambda x: split_category(x))

## 2023년에 구매한 제품카테고리

df_2023 = df2.loc[df2['년도'] == 2023]

grouped_data = df_2023.groupby('고객ID')['제품카테고리'].apply(list).reset_index()
grouped_data

df2 = df2.merge(grouped_data, on='고객ID', how = 'left')

df2 = df2.rename({'제품카테고리_x' : '제품카테고리'}, axis = 1)
df2 = df2.rename({'제품카테고리_y' : '제품리스트'}, axis = 1)

df2.head(1)

## 제품카테고리 별 label 컬럼 추가

# 제품명과 label 컬럼명 매핑
product_label_mapping = {
    '에어솔루션': 'label_에어솔루션',
    '청소': 'label_청소',
    '부엌': 'label_부엌',
    '디스플레이': 'label_디스플레이',
    '건강관리': 'label_건강관리',
    '기타': 'label_기타'
}

# 각 행에 대해 조건에 맞게 값을 할당
for index, row in tqdm(df2.iterrows()):
    if row['2023년구매'] == 0:
        for label in product_label_mapping.values():
            df2.loc[index, label] = 0
    elif row['2023년구매'] == 1:
        purchased_products = row['제품리스트']  # 예: ['디스플레이', '청소, 빨래', ...]
        for product in purchased_products:
            label = product_label_mapping.get(product, None)
            if label:
                df2.at[index, label] = 1

df2['label_에어솔루션'].fillna(0)
df2['label_청소'].fillna(0)
df2['label_부엌'].fillna(0)
df2['label_디스플레이'].fillna(0)
df2['label_기타'].fillna(0)

## 고객별 데이터 생성

col_list = [
'고객ID','멤버십고객여부', '케어솔루션품목수', '케어십가입여부', '케어십품목수', '최근3년구매금액', '최근구매채널',
'Only&Best품목수', '건강관리가전품목수', '연령대', '주거형태', '아파트시세', '아파트평형대', '포인트사용빈도',
'보유멤버십포인트', '멤버십가입경로', '제휴카드보유여부', '프리미엄카드보유여부', '제휴가입여부', '멤버십앱설치여부',
'거주광역명', '거주시군구명', '제품카테고리', 'label_에어솔루션', 'label_청소', 'label_부엌',
       'label_디스플레이', 'label_건강관리', 'label_기타'] 
all_df = final_df_maker(df2, col_list)

all_df

# df_list = [['air', '에어솔루션'], ['clean', '청소, 빨래'], ['kitchen', '부엌'], ['display', '디스플레이'], ['health', '건강관리'], ['etc', '기타']]
